gql: split queries from mutations
